clean up structure
